+++ CHANGELOG.md
@@ -10,13 +10,17 @@ and this project adheres to
 
 ### Added
 
-- [Google Shopping] Add new API.
-
 ### Changed
 
 ### Fixed
 
 ### Removed
+
+## [1.1.0] - 2023-02-02
+
+### Added
+
+- [Google Shopping] Add new API.
 
 ## [1.0.0] - 2023-01-19
 
@@ -143,7 +147,8 @@ The API signatures of all functions are as follows:
 - `getAccount(parameters?, callback?)`
 - `getLocations(parameters?, callback?)`
 
-[unreleased]: https://github.com/serpapi/serpapi-javascript/compare/1.0.0...HEAD
+[unreleased]: https://github.com/serpapi/serpapi-javascript/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/serpapi/serpapi-javascript/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/serpapi/serpapi-javascript/compare/0.0.13...1.0.0
 [0.0.13]: https://github.com/serpapi/serpapi-javascript/compare/0.0.12...0.0.13
 [0.0.12]: https://github.com/serpapi/serpapi-javascript/compare/0.0.11...0.0.12